generating file based on whitelist
